+++ examples/hello_world_1.py
@@ -25,7 +25,7 @@ head.add(meta := html5.Meta(charset='utf-8'))
 head.add(meta := html5.Meta(name='viewport', content='width=device-width'))
 head.add(link := html5.Link(rel='shortcut icon', type='image/png', href='/static/gladius/favicon.png'))
 head.add(title := html5.Title(content='Hello world 1'))
-head.add(link := html5.Link(href='https://cdn.jsdelivr.net/npm/daisyui@2.6.0/dist/full.css', rel='stylesheet', type='text/css'))
+head.add(link := html5.Link(href='https://cdn.jsdelivr.net/npm/daisyui@3.7.7/dist/full.css', rel='stylesheet', type='text/css'))
 head.add(script := html5.Script(src='https://cdn.tailwindcss.com'))
 head.add(script := html5.Script(src='https://unpkg.com/htmx.org@1.9.5'))
 head.add(script := html5.Script(src='https://unpkg.com/htmx.org/dist/ext/debug.js'))
@@ -34,7 +34,7 @@ head.add(script := html5.Script(src='https://unpkg.com/htmx.org/dist/ext/event-h
 head.add(script := html5.Script(src='/static/gladius/multi-path-deps.js'))
 
 # body
-html.add(body := html5.Body(hx_ext='multi-path-deps'))
+html.add(body := html5.Body(hx_ext='multi-path-deps').add_class('p-10'))
 body.add(vflex := d.VFlex())
 
 # top cards
@@ -47,9 +47,9 @@ card.add(world_text := d.Text('World'))
 # buttons
 vflex.add(hflex := d.Flex())
 hflex.add(join := d.Join())
-join.add(button := d.Button(class_='btn btn-primary', onclick=hello_button_click))
+join.add(button := d.Button(onclick=hello_button_click).add_class('btn-primary'))
 button.add(text := d.Text('Hello'))
-join.add(button := d.Button(class_='btn btn-success', onclick=world_button_click))
+join.add(button := d.Button(onclick=world_button_click).add_class('btn-secondary'))
 button.add(text := d.Text('World'))
 
 # router